backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
pydantic[email]==2.7.1
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.27.0
